add useSyncEditorValue util
